Time: 0 ms (100%), Space: 9.1 MB (82.54%) - LeetHub

+++ 0008-string-to-integer-atoi/0008-string-to-integer-atoi.cpp
@@ -1,7 +1,6 @@
 class Solution {
 public:
     int myAtoi(string s) {
-        // string result="";
         long long n=s.size();
         int i=0;
         while(i<n && s[i]==' '){
@@ -14,13 +13,11 @@ public:
             }
             i++;
         }
-        long long result=0;
+        long result=0;
         while (i < n && isdigit(s[i])) {
             int digit = s[i] - '0';
             
             result = result * 10 + digit;
-            
-            // // Check for 32-bit overflow / underflow
             if (sign * result >= INT_MAX) {
                 return INT_MAX;
             }
@@ -31,15 +28,6 @@ public:
             i++;
         }
         return sign*result;
-        // for(int i=0;i<n;i++){
-        //     if(s[i]==" "){
-        //         continue;
-        //     }
-        //     else if(s[i]=='-' || s[i]=='+'){
-        //         result.push_back(s[i]);
-        //     }
-        //     else if()
-        // }
         
     }
 };
